backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
langchain-groq==0.2.0
chromadb==0.5.11
sentence-transformers==3.1.1
pypdf==5.0.1
